requirements.txt
python-telegram-bot
python-binance
aiomysql
python-dotenv
